back-end/requirement.txt
click==8.0.1
dataclasses==0.8
dnspython==2.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-PyMongo==2.3.0
importlib-metadata==4.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
